refactor: separated api and svc layer for environments
