+++ api/environments/routes.go
@@ -1,13 +1,16 @@
 package environments
 
 import (
-	"github.com/envsecrets/envsecrets/internal/environments/commons"
 	"github.com/labstack/echo/v4"
+)
+
+const (
+	ENV_ID = "env_id"
 )
 
 func AddRoutes(sg *echo.Group) {
 
-	group := sg.Group("/environments/:" + commons.ENV_ID)
+	group := sg.Group("/environments/:" + ENV_ID)
 
 	group.POST("/sync-password", SyncWithPasswordHandler)
 	group.POST("/sync", SyncHandler)
