+++ internal/actions/handler.go
@@ -6,7 +6,6 @@ import (
 	"github.com/envsecrets/envsecrets/internal/clients"
 	"github.com/envsecrets/envsecrets/internal/context"
 	"github.com/envsecrets/envsecrets/internal/environments"
-	environmentCommons "github.com/envsecrets/envsecrets/internal/environments/commons"
 	"github.com/envsecrets/envsecrets/internal/projects"
 	"github.com/envsecrets/envsecrets/internal/subscriptions"
 	"github.com/labstack/echo/v4"
@@ -24,7 +23,7 @@ func EnvironmentCreate(c echo.Context) error {
 	}
 
 	//	Unmarshal the data interface to our required entity.
-	var options environmentCommons.CreateOptions
+	var options environments.CreateOptions
 	if err := MapToStruct(payload.Input.Args, &options); err != nil {
 		return c.JSON(http.StatusBadRequest, &clients.APIResponse{
 			Message: "failed to unmarshal new data",
