--- internal/environments/commons/commons.go
@@ -1,5 +1,0 @@
-package commons
-
-const (
-	ENV_ID = "env_id"
-)
