+++ internal/environments/dto.go
@@ -1,4 +1,4 @@
-package commons
+package environments
 
 import (
 	"time"
