+++ main.go
@@ -6,10 +6,10 @@ import (
 	"os"
 	"strings"
 
+	"github.com/envsecrets/envsecrets/api/environments"
 	"github.com/envsecrets/envsecrets/api/integrations"
 	"github.com/envsecrets/envsecrets/internal/actions"
 	"github.com/envsecrets/envsecrets/internal/auth"
-	"github.com/envsecrets/envsecrets/internal/environments"
 	"github.com/envsecrets/envsecrets/internal/events"
 	"github.com/envsecrets/envsecrets/internal/invites"
 	"github.com/envsecrets/envsecrets/internal/keys"
